Move `json_function_tests.go` and json tests that depend on it to their own package.
This ensures that non-test code in `sql/expression/function/json` doesn't depend on `testify`, which is a library that we only want to depend on for tests.
